+++ src/Core/MedicalCenters.Domain/Entities/IntermediateEntities/Doctor_Visit.cs
@@ -9,7 +9,7 @@ namespace MedicalCenters.Domain.Entities.IntermediateEntities
     public class Doctor_Visit : BaseCreateableDomainEntity
     {
         [Required]
-        public long DoctorId { get; set; }
+        public int DoctorId { get; set; }
         public Doctor Doctor { get; set; }
 
         [Required]
