+++ src/Core/MedicalCenters.Domain/Entities/Shifts/ShiftPlan.cs
@@ -13,7 +13,7 @@ namespace MedicalCenters.Domain.Entities.Shifts
         [Key]
         public long Id { get; set; }
         [Required]
-        public long PersonelId { get; set; }
+        public int PersonelId { get; set; }
         public Personnel Personel { get; set; }
         [Required]
         public int MedicalUnitId { get; set; }
